feat: add unit tests for scrape function
Adds unit tests for the scrape function, which covers the extraction of
production, inverter, and battery stats. Refactors the code to make it
more testable by introducing an EnvoyClient interface and a mock
implementation for testing.

Removes the CommCheck function from the scrape function for now, as there
are issues with the go-envoy library that prevent it from being tested.